+++ javaparser-core/src/main/java/com/github/javaparser/ast/validator/Java5Validator.java
@@ -1,8 +1,11 @@
 package com.github.javaparser.ast.validator;
 
+import com.github.javaparser.ast.Modifier;
 import com.github.javaparser.ast.Node;
 import com.github.javaparser.ast.NodeList;
+import com.github.javaparser.ast.expr.VariableDeclarationExpr;
 import com.github.javaparser.ast.nodeTypes.NodeWithTypeArguments;
+import com.github.javaparser.ast.stmt.ForEachStmt;
 import com.github.javaparser.ast.type.PrimitiveType;
 import com.github.javaparser.ast.type.Type;
 
@@ -32,6 +35,17 @@ public class Java5Validator extends Java1_4Validator {
         }
     });
 
+    // Enhanced for statements were introduced in Java 5. There must be exactly one declared variable, and the only
+    // allowed modifier is FINAL.
+    Validator forEachStmt = new SingleNodeTypeValidator<>(ForEachStmt.class, (node, reporter) -> {
+        VariableDeclarationExpr declaration = node.getVariable();
+        // assert that the variable declaration expression has exactly one variable declarator
+        if (declaration.getVariables().size() != 1) {
+            reporter.report(node, "A foreach statement's variable declaration must have exactly one variable " +
+                                  "declarator. Given: " + declaration.getVariables().size() + ".");
+        }
+    });
+
     protected final Validator enumNotAllowed = new ReservedKeywordValidator("enum");
 
     public Java5Validator() {
@@ -39,7 +53,8 @@ public class Java5Validator extends Java1_4Validator {
         replace(noGenerics, genericsWithoutDiamondOperator);
         add(noPrimitiveGenericArguments);
         add(enumNotAllowed);
-        
+        add(forEachStmt);
+
         // TODO validate annotations on classes, fields and methods but nowhere else
         // The following is probably too simple.
         remove(noAnnotations);