+++ javaparser-core/src/main/java/com/github/javaparser/ast/visitor/GenericListVisitorAdapter.java
@@ -1484,11 +1484,11 @@ public abstract class GenericListVisitorAdapter<R, A> implements GenericVisitor<
         return result;
     }
 
-    public List<R> visit(final SwitchEntryStmt n, final A arg) {
+    public List<R> visit(final SwitchEntry n, final A arg) {
         List<R> result = new ArrayList<>();
         List<R> tmp;
-        if (n.getLabel().isPresent()) {
-            tmp = n.getLabel().get().accept(this, arg);
+        {
+            tmp = n.getLabels().accept(this, arg);
             if (tmp != null)
                 result.addAll(tmp);
         }