+++ javaparser-core/src/main/java/com/github/javaparser/ast/visitor/VoidVisitorAdapter.java
@@ -493,8 +493,8 @@ public abstract class VoidVisitorAdapter<A> implements VoidVisitor<A> {
     }
 
     @Override
-    public void visit(final SwitchEntryStmt n, final A arg) {
-        n.getLabel().ifPresent(l -> l.accept(this, arg));
+    public void visit(final SwitchEntry n, final A arg) {
+        n.getLabels().forEach(p -> p.accept(this, arg));
         n.getStatements().forEach(p -> p.accept(this, arg));
         n.getComment().ifPresent(l -> l.accept(this, arg));
     }